Remove thisContext, as it creates dangerous ambient authority.
Activation mirrors and exceptions should have access to the current context via a VM mirror capability and attenuate it. In the current Squeak implementation, exceptions are still implemented by a Smalltalk class that has access to thisContext. In Primordial Soup, exceptions access the current context through a primitive only accessible to the platform sources.