Modified build script, added deploy scripts. Changed the way the PWA version numbers are managed and stored in git.
Only the PWA version for newspeaklanguage.org is kept in git. Others, such as the PWA version for localhost or any other URL is not kept in git, but users can choose to manage it locally outside of git.

README.md is significantly modified and expanded, it now describes more details about
installing Newspeak from scratch, building, deploying and serving Newspeak applications.