WIP. Refinements to the built-in help in the IDE. The end goal is to mak ethe IDE self-documenting from the user point of view. One should be able to earn how to use the system via the help buttons, from the top of the IDE window, down thru every piece of UI.
This should obviate the need for a separate IDE tutorial like the out-of-date "NS101: A Guide to the Perplexed". Separate language tutorials etc. can be integrated into this scheme as well - say by placing them in documents referenced by the help. These dcos could be in the IDE namespace or online.

This documentation should ultimately serve both humans and machines. An AI should be able to  learn how to drive the system interactively as well - though a programmatic API is even more useful.

At this point, there are improvements to the help in the IDE window/toolbar, the home page, namespaces and their entries, and nested class entries.