Remove Electron IDE from build script
Electron is being removed from the Ministry of Truth. With the latest changes to file saving, the web version is as convenient to use as the electron version. We have not updasted the electron version in a long time, we don't have the bandwidth, and the old version raises security concerns.